Login & logout pages with sessions
I made login & logout pages with sessions, as well as a very validation system for logging in (at the moment, only username=test and password=1234 will be valid). There is also a test page that the login page links to that should display the user's name if the user is logged in, and tell the user they're not logged in if they're not. You can also test this on my CODD server: codd.cs.gsu.edu/~jkouttu1/Project/01/login.php

Still need to figure out databases if we're going that route.